+++ src/drawableObjects/Line.java
@@ -1,13 +1,13 @@
 package drawableObjects;
 
 public class Line{
-	private float xCoordStart;
-	private float yCoordStart;
-	private float xCoordEnd;
-	private float yCoordEnd;
+	private int xCoordStart;
+	private int yCoordStart;
+	private int xCoordEnd;
+	private int yCoordEnd;
 	private int color;
 	
-	public Line(float iXCoordStart,float iYCoordStart,float iXCoodEnd,float iYCoordEnd,int iColor){
+	public Line(int iXCoordStart,int iYCoordStart,int iXCoodEnd,int iYCoordEnd,int iColor){
 		xCoordStart = iXCoordStart;
 		xCoordEnd = iXCoodEnd;
 		yCoordStart = iYCoordStart;
@@ -15,7 +15,7 @@ public class Line{
 		color=iColor;
 	}
 
-	public Line(float iXCoordStart,float iYCoordStart,float iXCoodEnd,float iYCoordEnd) {
+	public Line(int iXCoordStart,int iYCoordStart,int iXCoodEnd,int iYCoordEnd) {
 		xCoordStart = iXCoordStart;
 		xCoordEnd = iXCoodEnd;
 		yCoordStart = iYCoordStart;
@@ -23,35 +23,35 @@ public class Line{
 		color=255;
 	}
 
-	public void setXCoordStart(float iXCoordStart) {
+	public void setXCoordStart(int iXCoordStart) {
 		xCoordStart = iXCoordStart;
 	}
 
-	public float getXCoordStart() {
+	public int getXCoordStart() {
 		return xCoordStart;
 	}
 
-	public void setYCoordStart(float iYCoordStart) {
+	public void setYCoordStart(int iYCoordStart) {
 		yCoordStart = iYCoordStart;
 	}
 
-	public float getYCoordStart() {
+	public int getYCoordStart() {
 		return yCoordStart;
 	}
 
-	public void setXCoordEnd(float iXCoordEnd) {
+	public void setXCoordEnd(int iXCoordEnd) {
 		xCoordEnd = iXCoordEnd;
 	}
 
-	public float getXCoordEnd() {
+	public int getXCoordEnd() {
 		return xCoordEnd;
 	}
 
-	public void setYCoordEnd(float iYCoordEnd) {
+	public void setYCoordEnd(int iYCoordEnd) {
 		yCoordEnd = iYCoordEnd;
 	}
 
-	public float getYCoordEnd() {
+	public int getYCoordEnd() {
 		return yCoordEnd;
 	}
